upgrade fingers chart using vue3-apexchart
